requirements.txt
jsonschema<2.6
PyYAML==3.11
